add loading to new event modal to prevent bug when throttling
