+++ client/src/components/LeftDrawer.jsx
@@ -1,7 +1,8 @@
 import NewEventModal from './NewEventModal';
 import LoginButton from './LoginButton';
-import { useContext, useState } from 'react';
+import { useContext, useState, useEffect } from 'react';
 import { UserContext } from './App';
+import latLngToAddress from '../utilities/geocode';
 
 export default function LeftDrawer({
   point,
@@ -10,7 +11,24 @@ export default function LeftDrawer({
   fetchEvents,
 }) {
   const currUser = useContext(UserContext);
-  const [address, setAddress] = useState('Address not set.');
+  const [address, setAddress] = useState(false);
+  const [loadingAddress, setLoadingAddress] = useState(true)
+
+  async function getAddress() {
+    try {
+      setLoadingAddress(true)
+      setAddress(false)
+      const geocodeLatLon = await latLngToAddress(point[0], point[1]);
+      setAddress(geocodeLatLon)
+      setLoadingAddress(false)
+    } catch (err) {
+      console.log(err);
+    }
+  }
+
+  useEffect(() => {
+    getAddress();
+  }, [point]);
 
   return (
     <div className='w-full sm:w-3/5 flex flex-col justify-center items-center z-10 bg-white absolute bottom-0 border-t-1 border-primary left-auto'>
@@ -22,17 +40,27 @@ export default function LeftDrawer({
             </div>
             <div className='w-full'>
               <div>
+              {!loadingAddress ? (
+                <div >
                 {address ? (
                   <div className='h-1/6 w-3/5'>
                     <h2>Nice! You've found a spot:</h2>
-                    <p className='min-w-full'>{address.name}.</p>
+                    <p className='min-w-full'>{address.name}</p>
                   </div>
                 ) : (
                   <div className='h-1/6 w-3/5'>
-                    <h2>We can't find an address at that location.</h2>
+                    <h2>We can't find an address there.</h2>
                     <h2>Please find a different spot.</h2>
                   </div>
                 )}
+              </div>
+              ):(
+                <div className='h-1/6 w-3/5'>
+                    <h2>Looking for an address...</h2>
+                  </div>
+              )}      
+
+
               </div>
               <div className='mt-5'>
                 {currUser ? (
